% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_function_table.R
\name{add_inverse_function}
\alias{add_inverse_function}
\title{Add Inverse Function Mapping (Mutable Table)}
\usage{
add_inverse_function(fname, inv_fname, table = inverse_function_table)
}
\arguments{
\item{fname}{Character. Name of the function.}

\item{inv_fname}{Character. Name of the inverse function.}

\item{table}{Data frame. Registry of function-inverse pairs. (default = inverse_function_table)}
}
\value{
The updated table, invisibly.
}
\description{
Add Inverse Function Mapping (Mutable Table)
}
