% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_function_table.R
\name{get_inverse_function}
\alias{get_inverse_function}
\title{Get the Inverse Function Name}
\usage{
get_inverse_function(fname, table = NULL)
}
\arguments{
\item{fname}{Character. Function name.}
}
\value{
Character vector of the inverse function name(s), or NULL if not found.
}
\description{
Retrieves the inverse function name for a given function name, searching both directions.
}
\examples{
get_inverse_function("log")

}
