% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_version_desc.R
\name{increment_version_desc}
\alias{increment_version_desc}
\title{Increment Package Version in DESCRIPTION Using desc}
\usage{
increment_version_desc(
  folder = ".",
  increase = c("patch", "minor", "major", "build")
)
}
\arguments{
\item{folder}{Package folder containing DESCRIPTION.}

\item{increase}{Character. One of "major", "minor", "patch", "build".}
}
\value{
Named character vector with old and new version.
}
\description{
Increments the version in DESCRIPTION by "major", "minor", "patch", or "build".
Here, "build" is translated to "dev".
}
