% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_function_table.R
\docType{data}
\name{inverse_function_table}
\alias{inverse_function_table}
\title{Table of Inverse Functions}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\usage{
inverse_function_table
}
\description{
Holds the mapping of functions to their inverses by function name (as character strings, not functions).
}
\details{
The table can be queried or expanded with new function-inverse pairs.
}
\keyword{datasets}
