% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_na_helpers.R
\name{na_false}
\alias{na_false}
\title{na_false}
\usage{
na_false(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A logical vector where any NA is coerced to FALSE.
}
\description{
Treat NA values in a logical vector as FALSE.
}
\examples{
na_false(c(NA, FALSE, TRUE))
}
