% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_na_helpers.R
\name{na_true}
\alias{na_true}
\title{na_true}
\usage{
na_true(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A logical vector where any NA is coerced to TRUE.
}
\description{
Treat NA values in a logical vector as TRUE.
}
\examples{
na_true(c(NA, FALSE, TRUE))
}
