% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_difference_statistics.R
\name{print.difference_statistic}
\alias{print.difference_statistic}
\title{Print method for difference_statistic objects}
\usage{
\method{print}{difference_statistic}(x, ...)
}
\arguments{
\item{x}{A difference_statistic object}

\item{...}{ignored}
}
\description{
Print method for difference_statistic objects
}
