% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_function_table.R
\name{reciprocal}
\alias{reciprocal}
\title{reciprocal}
\usage{
reciprocal(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of reciprocals.
}
\description{
Computes the reciprocal (1/x) of its argument.
}
\examples{
reciprocal(c(2, 4))

}
