% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Replace Values in a Vector or Factor}
\usage{
replace_values(x, old_values, new_values, warn = TRUE, drop_levels = TRUE)
}
\arguments{
\item{x}{A vector or factor whose values will be replaced.}

\item{old_values}{A vector of values to be replaced. Must not contain duplicates.}

\item{new_values}{A vector of replacement values. Must be the same length as \code{old_values} or of length 1.}

\item{warn}{Logical, if TRUE, warns about unmatched old values. Default: TRUE.}

\item{drop_levels}{Logical, if TRUE and \code{x} is a factor, other unused levels are dropped after replacement. Default: TRUE.}
}
\value{
Returns \code{x} with specified values replaced by new values. If \code{x} is a factor, factor levels are updated accordingly.
}
\description{
Replaces one or more values in a vector or factor, including levels of factors.
}
\examples{
replace_values(c(1, 2, 2, 3), c(2, 3), c(99, 42))
replace_values(factor(c("a", "b", "b", "c")), c("b", "c"), c("x", "y"))
replace_values(letters, "a", "z")

}
