% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_function_table.R
\name{sqr}
\alias{sqr}
\title{sqr}
\usage{
sqr(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of squared values.
}
\description{
Computes the square of its argument.
}
\examples{
sqr(c(3, 4))

}
