% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_difference_statistics.R
\name{summary.difference_statistic}
\alias{summary.difference_statistic}
\title{Summary method for difference_statistic objects}
\usage{
\method{summary}{difference_statistic}(object, sig.level = 0.05, ...)
}
\arguments{
\item{object}{A difference_statistic object}

\item{sig.level}{Significance level for counting significant results}

\item{...}{ignored}
}
\description{
Summary method for difference_statistic objects
}
